WEA-6 Refactor Code for progressDialog and Titlebar, Change package main to home
