+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
       android:supportsRtl="true"
       android:theme="@style/AppTheme">
     <activity
-        android:name=".main.MainActivity"
+        android:name=".home.HomeActivity"
         android:screenOrientation="portrait"
         android:theme="@style/SplashTheme">
       <intent-filter>
