--- app/src/main/java/com/ayush/weatherapp/BasePresenter.java
@@ -1,8 +1,0 @@
-package com.ayush.weatherapp;
-
-import android.app.ProgressDialog;
-
-public interface BasePresenter {
-
-}
-
