+++ app/src/main/java/com/ayush/weatherapp/home/HomeActivity.java
@@ -1,6 +1,5 @@
-package com.ayush.weatherapp.main;
+package com.ayush.weatherapp.home;
 
-import android.app.ProgressDialog;
 import android.os.Bundle;
 import android.support.design.widget.NavigationView;
 import android.support.v4.view.GravityCompat;
@@ -11,23 +10,20 @@ import android.view.MenuItem;
 import android.widget.TextView;
 import android.widget.Toast;
 import butterknife.BindView;
-import com.ayush.weatherapp.BaseActivity;
 import com.ayush.weatherapp.R;
+import com.ayush.weatherapp.mvp.BaseActivity;
 
-public class MainActivity extends BaseActivity implements MainContract.View {
+public class HomeActivity extends BaseActivity implements HomeContract.View {
 
   @BindView(R.id.layout_drawer) DrawerLayout drawerLayout;
   @BindView(R.id.nav_view) NavigationView navigationView;
   @BindView(R.id.txt_main) TextView txtCurrentForecastSummary;
+  @BindView(R.id.toolbar) android.support.v7.widget.Toolbar toolbar;
 
-  MainContract.Presenter presenter;
+  HomeContract.Presenter presenter;
 
   @Override protected int getContextView() {
     return R.layout.navigation_drawer;
-  }
-
-  @Override protected int getToolbarResource() {
-    return R.id.toolbar;
   }
 
   @Override protected void onCreate(Bundle savedInstanceState) {
@@ -36,12 +32,12 @@ public class MainActivity extends BaseActivity implements MainContract.View {
 
     super.onCreate(savedInstanceState);
 
+    initToolbar(toolbar);
+    showTitleBar(false);
+
     setNavigationView();
 
-    getCurrentActionBar().setDisplayShowTitleEnabled(false);
-
-    presenter = new MainPresenter(this);
-    presenter.setView(this);
+    presenter = new HomePresenterImpl(this, this );
 
     presenter.fetchWeatherDetails();
   }
@@ -64,21 +60,26 @@ public class MainActivity extends BaseActivity implements MainContract.View {
   private void setNavigationView() {
     navigationView.setNavigationItemSelectedListener(menuItem -> {
       //TODO
-      Toast.makeText(MainActivity.this, menuItem.toString(), Toast.LENGTH_SHORT).show();
+      Toast.makeText(HomeActivity.this, menuItem.toString(), Toast.LENGTH_SHORT).show();
 
       drawerLayout.closeDrawers();
-      
+
       // Add code here to update the UI based on the item selected
       // For example, swap UI fragments here
       return true;
     });
   }
 
-  @Override public ProgressDialog getProgressDialog() {
-    return progressDialog;
+  @Override public void showProgressDialog(String message) {
+    getProgressDialog().setMessage(message);
+    getProgressDialog().show();
   }
 
-  @Override public void setCurrentTemperatureSummary(String s) {
-    txtCurrentForecastSummary.setText(s);
+  @Override public void hideProgressDialog() {
+    getProgressDialog().dismiss();
+  }
+
+  @Override public void setCurrentTemperatureSummary(String summary) {
+    txtCurrentForecastSummary.setText(summary);
   }
 }
