+++ app/src/main/java/com/ayush/weatherapp/mvp/BaseContract.java
@@ -1,0 +1,12 @@
+package com.ayush.weatherapp.mvp;
+
+public interface BaseContract {
+  interface View {
+    void showProgressDialog(String message);
+
+    void hideProgressDialog();
+  }
+
+  interface Presenter {
+  }
+}
